+++ src/destroyer.lua
@@ -3,6 +3,7 @@ local Container = Addon.Container
 local Destroyer = Addon.Destroyer
 local E = Addon.Events
 local EventManager = Addon.EventManager
+local Items = Addon.Items
 local JunkFilter = Addon.JunkFilter
 local L = Addon.Locale
 
@@ -15,6 +16,17 @@ local function sortByTotalPrice(a, b)
   local aTotalPrice = (a.price * a.quantity)
   local bTotalPrice = (b.price * b.quantity)
   return aTotalPrice == bTotalPrice and a.quality > b.quality or aTotalPrice > bTotalPrice
+end
+
+local function handleItem(item)
+  if not Items:IsItemStillInBags(item) then return end
+  if Items:IsItemLocked(item) then return end
+
+  ClearCursor()
+  Container.PickupContainerItem(item.bag, item.slot)
+  DeleteCursorItem()
+
+  EventManager:Fire(E.AttemptedToDestroyItem, item)
 end
 
 -- ============================================================================
@@ -33,10 +45,10 @@ function Destroyer:Start()
   table.sort(items, sortByTotalPrice)
   local item = table.remove(items)
 
-  -- Delete item.
-  ClearCursor()
-  Container.PickupContainerItem(item.bag, item.slot)
-  DeleteCursorItem()
+  -- Handle item.
+  handleItem(item)
+end
 
-  EventManager:Fire(E.AttemptedToDestroyItem, item)
+function Destroyer:HandleItem(item)
+  if not Addon:IsBusy() then handleItem(item) end
 end