+++ src/ui/merchant-button.lua
@@ -5,7 +5,6 @@ local E = Addon.Events
 local EventManager = Addon.EventManager
 local L = Addon.Locale
 local SavedVariables = Addon.SavedVariables
-local Sounds = Addon.Sounds
 local Tooltip = Addon.Tooltip
 
 EventManager:Once(E.SavedVariablesReady, function()
@@ -13,7 +12,7 @@ EventManager:Once(E.SavedVariablesReady, function()
   frame:SetText(ADDON_NAME)
   frame:SetWidth(frame:GetTextWidth() + 32)
   frame:SetHeight(frame:GetTextHeight() + 12)
-  frame:SetPoint("TOPRIGHT", MerchantFrame, "BOTTOMRIGHT", 0, 0)
+  frame:SetPoint("TOPLEFT", 60, -28)
   frame:RegisterForClicks("LeftButtonUp", "RightButtonUp")
 
   MerchantFrame:HookScript("OnUpdate", function()